+++ src/lib.rs
@@ -4,62 +4,20 @@ extern crate rand;
 pub mod api;
 pub mod array_ops;
 pub mod categorical;
+pub mod data;
 pub mod iter_mean;
 pub mod vec2d;
 
 
 use std::marker::PhantomData;
 use rand::{thread_rng, Rng};
+
+use data::{SampleDescription, TrainingData};
 use iter_mean::IterMean;
 
 pub struct Split<Theta, Threshold> {
     pub theta: Theta,
     pub threshold: Threshold,
-}
-
-/// Trait for a dataset that can be used for training a decision tree
-pub trait TrainingData<Sample>
-    where Sample: SampleDescription
-{
-    /// Return number of samples in the data set
-    fn n_samples(&self) -> usize;
-
-    /// Generate a new split feature (typically, this will be randomized)
-    fn gen_split_feature(&self) -> Sample::ThetaSplit;
-
-    /// Train a new leaf predictor
-    fn train_leaf_predictor(&self) -> Sample::ThetaLeaf;
-
-    /// Partition data set in-place according to a split
-    fn partition_data(&mut self, split: &Split<Sample::ThetaSplit, Sample::Feature>) -> (&mut Self, &mut Self);
-
-    /// Compute split criterion
-    fn split_criterion(&self) -> f64;
-
-    /// Return minimum and maximum value of a feature
-    fn feature_bounds(&self, theta: &Sample::ThetaSplit) -> (Sample::Feature, Sample::Feature);
-}
-
-/// Trait that describes a Sample used with decision trees
-pub trait SampleDescription {
-    /// Type used to parametrize split features
-    type ThetaSplit;
-
-    /// Type used to parametrize leaf predictors
-    type ThetaLeaf;
-
-    /// Type of a split feature
-    type Feature: PartialOrd + rand::distributions::range::SampleRange;
-
-    /// Type of predicted values; this can be the same as `Self::Y` (e.g. regression) or something
-    /// different (e.g. class probabilities).
-    type Prediction;
-
-    /// Compute the value of a leaf feature for a given sample
-    fn sample_as_split_feature(&self, theta: &Self::ThetaSplit) -> Self::Feature;
-
-    /// Compute the leaf prediction for a given sample
-    fn sample_predict(&self, w: &Self::ThetaLeaf) -> Self::Prediction;
 }
 
 /// Find split
@@ -71,7 +29,9 @@ pub trait SplitFinder
         -> Option<Split<Sample::ThetaSplit, Sample::Feature>>;
 }
 
-/// A decision tree node. Can be either a split node with a Splitter and two children, or a leaf
+/// A decision tree node.
+///
+/// Can be either a split node with a Splitter and two children, or a leaf
 /// node with a LeafPredictor.
 pub enum Node<T>
     where T: SampleDescription
